requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.4.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
flake8==4.0.1
fonttools==4.29.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.0
openpyxl==3.0.9
oscrypto==1.2.1
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pure-eval==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.12.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
selenium==4.1.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.2
snowflake-sqlalchemy==1.3.3
sortedcontainers==2.4.0
splinter==0.17.0
SQLAlchemy==1.4.29
stack-data==0.1.3
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wsproto==1.0.0
XlsxWriter==3.0.2
